+++ db/migrate/20110215143010_add_timestamps_to_custom_fields.rb
@@ -1,9 +1,0 @@
-class AddTimestampsToCustomFields < ActiveRecord::Migration
-  def self.up
-    add_timestamps :custom_fields
-  end
-
-  def self.down
-    remove_timestamps :custom_fields
-  end
-end